travis yml mods and package updates
Was manually editing yml config and triggering builds for travis on their site, found that I needed to be compiling my TS (duh, dist is not committed). Additionally, @types/Underscore was throwing an error that stopped travis, even though it doesn't stop me. Updating the packages seems to have fixed that.